# api/openapi.yaml
openapi: 3.0.3
info:
  title: BeeBase Notification Service API
  description: >
    Push-device lifecycle service for BeeBase. Access tokens are issued by
    auth-service; this service uses only the verified authenticated user ID
    and never trusts a user ID from a request body.
  version: 0.2.0
servers:
  - url: http://localhost:8080
tags:
  - name: system
    description: Liveness and readiness probes.
  - name: devices
    description: Register, update, and unregister push destinations.
  - name: reminders
    description: Authenticated, user-owned scheduled reminders.
security:
  - bearerAuth: []
paths:
  /health:
    get:
      tags: [system]
      summary: Liveness probe
      security: []
      responses:
        "200":
          description: The process is alive.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StatusResponse" }
  /ready:
    get:
      tags: [system]
      summary: Readiness probe
      security: []
      responses:
        "200":
          description: The service and database are ready.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StatusResponse" }
        "503":
          description: The database is unreachable.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StatusResponse" }
  /api/v1/devices:
    post:
      tags: [devices]
      summary: Register or refresh the authenticated user's current device
      description: >
        Registers the destination explicitly after authentication or startup.
        Repeating the request for the same destination and user is idempotent.
        A user may register multiple devices. The destination is a Firebase
        Installation ID (FID), not an access token and not a user ID.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DeviceRequest" }
      responses:
        "200":
          description: The device was registered or refreshed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PushDevice" }
        "400":
          description: The request failed validation.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Missing, invalid, or expired access token.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "409":
          description: The destination is currently owned by another user.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /api/v1/devices/{deviceID}:
    parameters:
      - $ref: "#/components/parameters/DeviceID"
    put:
      tags: [devices]
      summary: Update a registered device destination
      description: >
        Replaces the destination and/or platform for a device owned by the
        authenticated user. Flutter should use this when Firebase rotates or
        changes the current installation ID.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DeviceRequest" }
      responses:
        "200":
          description: The device was updated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PushDevice" }
        "400":
          description: The request failed validation.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Missing, invalid, or expired access token.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: The device does not belong to the authenticated user.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "409":
          description: The replacement destination is owned by another user.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    delete:
      tags: [devices]
      summary: Unregister a device
      description: >
        Removes the device owned by the authenticated user. Flutter should
        call this on logout or when the user opts out of push notifications.
      responses:
        "204": { description: The device was removed. }
        "401":
          description: Missing, invalid, or expired access token.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: The device does not belong to the authenticated user.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /api/v1/reminders:
    get:
      tags: [reminders]
      summary: List the authenticated user's reminders
      parameters:
        - {name: entity_type, in: query, schema: {type: string, enum: [apiary, hive, inspection, harvest]}}
        - {name: entity_id, in: query, schema: {type: string, format: uuid}}
        - {name: status, in: query, schema: {type: string, enum: [scheduled, processing, sent, cancelled, failed]}}
        - {name: page, in: query, schema: {type: integer, minimum: 1}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100}}
      responses:
        "200": {description: Paginated reminders., content: {application/json: {schema: {$ref: "#/components/schemas/ReminderPage"}}}}
    post:
      tags: [reminders]
      summary: Create a custom reminder
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/ReminderRequest"}}}}
      responses:
        "201": {description: Created., content: {application/json: {schema: {$ref: "#/components/schemas/Reminder"}}}}
  /api/v1/reminders/{reminderID}:
    parameters: [{name: reminderID, in: path, required: true, schema: {type: string, format: uuid}}]
    get:
      tags: [reminders]
      summary: Get an owned reminder
      responses: {"200": {description: Reminder., content: {application/json: {schema: {$ref: "#/components/schemas/Reminder"}}}}, "404": {description: Not found.}}
    put:
      tags: [reminders]
      summary: Update an owned reminder
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/ReminderRequest"}}}}
      responses: {"200": {description: Updated., content: {application/json: {schema: {$ref: "#/components/schemas/Reminder"}}}}}
    delete:
      tags: [reminders]
      summary: Delete an owned reminder
      responses: {"204": {description: Deleted.}, "404": {description: Not found.}}
components:
  parameters:
    DeviceID:
      name: deviceID
      in: path
      required: true
      description: The device UUID returned by registration.
      schema: { type: string, format: uuid }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    StatusResponse:
      type: object
      required: [status]
      properties:
        status: { type: string, example: ok }
    DeviceRequest:
      type: object
      required: [destination, platform]
      properties:
        destination:
          type: string
          minLength: 1
          description: Firebase Installation ID (FID) for the current app installation.
        platform:
          type: string
          enum: [ios, android]
    PushDevice:
      type: object
      required: [id, user_id, destination, platform, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string, format: uuid, description: Authenticated owner of the device. }
        destination: { type: string, description: Firebase Installation ID (FID). }
        platform: { type: string, enum: [ios, android] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    ReminderRequest:
      type: object
      required: [title, entity_type, entity_id, remind_at]
      properties:
        title: {type: string, minLength: 1, maxLength: 200}
        note: {type: string, maxLength: 2000}
        entity_type: {type: string, enum: [apiary, hive, inspection, harvest]}
        entity_id: {type: string, format: uuid}
        remind_at: {type: string, format: date-time, description: UTC instant; offsets are normalized to UTC.}
    Reminder:
      allOf:
        - $ref: "#/components/schemas/ReminderRequest"
        - type: object
          required: [id, user_id, reminder_type, source, status, created_at, updated_at]
          properties:
            id: {type: string, format: uuid}
            user_id: {type: string, format: uuid}
            reminder_type: {type: string, enum: [custom]}
            source: {type: string, enum: [manual, system]}
            status: {type: string, enum: [scheduled, processing, sent, cancelled, failed]}
            cancel_reason: {type: string, nullable: true}
            attempt_count: {type: integer}
            next_attempt_at: {type: string, format: date-time}
            last_error: {type: string, nullable: true}
            created_at: {type: string, format: date-time}
            updated_at: {type: string, format: date-time}
    ReminderPage:
      type: object
      required: [items, pagination]
      properties:
        items: {type: array, items: {$ref: "#/components/schemas/Reminder"}}
        pagination: {type: object}
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - missing_authorization
                - invalid_access_token
                - invalid_request
                - invalid_device
                - invalid_device_id
                - destination_owned
                - internal_error
            message: { type: string, example: request validation failed }